test: add Postgres E2E tests (#2005)
