+++ packages/amplify-e2e-tests/src/__tests__/rds-mysql-custom-claims-refersto-auth.test.ts
@@ -11,6 +11,7 @@ import {
   setupRDSInstanceAndData,
   sleep,
   updateAuthAddUserGroups,
+  getProjectMeta,
 } from 'amplify-category-api-e2e-core';
 import { existsSync, writeFileSync, removeSync } from 'fs-extra';
 import generator from 'generate-password';
@@ -34,7 +35,7 @@ describe('RDS userpool provider with custom Auth claims tests', () => {
   // Generate settings for RDS instance
   const username = db_user;
   const password = db_password;
-  const region = 'ap-northeast-2';
+  let region = 'us-east-1';
   let port = 3306;
   const database = 'default_db';
   let host = 'localhost';
@@ -91,6 +92,10 @@ describe('RDS userpool provider with custom Auth claims tests', () => {
       disableAmplifyAppCreation: false,
       name: projName,
     });
+
+    const metaAfterInit = getProjectMeta(projRoot);
+    region = metaAfterInit.providers.awscloudformation.Region;
+
     await addAuthWithPreTokenGenerationTrigger(projRoot);
     updatePreAuthTrigger(projRoot, 'user_id');
