+++ packages/amplify-e2e-tests/src/__tests__/rds-mysql-multi-auth-1.test.ts
@@ -10,23 +10,21 @@ import {
   setupRDSInstanceAndData,
   sleep,
   updateAuthAddUserGroups,
+  getProjectMeta,
 } from 'amplify-category-api-e2e-core';
 import { existsSync, writeFileSync, removeSync } from 'fs-extra';
 import generator from 'generate-password';
 import path from 'path';
 import { schema, sqlCreateStatements } from './auth-test-schemas/userpool-apikey';
 import {
-  createModelOperationHelpers,
   configureAppSyncClients,
   checkOperationResult,
   checkListItemExistence,
   appendAmplifyInput,
   getAppSyncEndpoint,
 } from '../rds-v2-test-utils';
-import { setupUser, getUserPoolId, signInUser, configureAmplify, getConfiguredAppsyncClientCognitoAuth } from '../schema-api-directives';
-import { gql } from 'graphql-tag';
+import { setupUser, getUserPoolId, signInUser, configureAmplify } from '../schema-api-directives';
 import { GQLQueryHelper } from '../query-utils/gql-helper';
-import { check } from 'yargs';
 
 // to deal with bug in cognito-identity-js
 (global as any).fetch = require('node-fetch');
@@ -37,7 +35,7 @@ describe('RDS Cognito userpool provider Auth tests', () => {
   // Generate settings for RDS instance
   const username = db_user;
   const password = db_password;
-  const region = 'ap-northeast-2';
+  let region = 'us-east-1';
   let port = 3306;
   const database = 'default_db';
   let host = 'localhost';
@@ -99,6 +97,9 @@ describe('RDS Cognito userpool provider Auth tests', () => {
       name: projName,
     });
 
+    const metaAfterInit = getProjectMeta(projRoot);
+    region = metaAfterInit.providers.awscloudformation.Region;
+
     await addApi(projRoot, {
       transformerVersion: 2,
       'Amazon Cognito User Pool': {},
@@ -115,13 +116,13 @@ describe('RDS Cognito userpool provider Auth tests', () => {
       port,
       username,
       password,
-      useVpc: false,
+      useVpc: true,
       apiExists: true,
     });
     writeFileSync(rdsSchemaFilePath, appendAmplifyInput(schema, 'mysql'), 'utf8');
 
     await updateAuthAddUserGroups(projRoot, [adminGroupName, devGroupName, moderatorGroupName]);
-    await amplifyPush(projRoot, true);
+    await amplifyPush(projRoot);
     await sleep(2 * 60 * 1000); // Wait for 2 minutes for the VPC endpoints to be live.
 
     const userPoolId = getUserPoolId(projRoot);
