+++ packages/amplify-e2e-tests/src/__tests__/rds-mysql-oidc-auth.test.ts
@@ -12,6 +12,7 @@ import {
   sleep,
   updateAuthAddUserGroups,
   amplifyPushWithoutCodegen,
+  getProjectMeta,
 } from 'amplify-category-api-e2e-core';
 import { existsSync, writeFileSync, removeSync } from 'fs-extra';
 import generator from 'generate-password';
@@ -46,7 +47,7 @@ describe('RDS OIDC provider Auth tests', () => {
   // Generate settings for RDS instance
   const username = db_user;
   const password = db_password;
-  const region = 'ap-northeast-2';
+  let region = 'us-east-1';
   let port = 3306;
   const database = 'default_db';
   let host = 'localhost';
@@ -116,6 +117,10 @@ describe('RDS OIDC provider Auth tests', () => {
       disableAmplifyAppCreation: false,
       name: projName,
     });
+
+    const metaAfterInit = getProjectMeta(projRoot);
+    region = metaAfterInit.providers.awscloudformation.Region;
+
     await addAuthWithPreTokenGenerationTrigger(projRoot);
     updatePreAuthTrigger(projRoot, 'user_id');
     await amplifyPushWithoutCodegen(projRoot);
